\name{mastPlot}
\alias{mastPlot}
\title{Plot mast model}
\description{
Plots data fitted with \code{\link{mastif}} in package \code{mastif}.
}
\usage{
  mastPlot(output, plotPars = NULL)
  
}
\arguments{
  \item{output}{\code{list} of class \code{mast}.}
  
  \item{plotPars}{\code{list} of inputs for plotting.}
}

\details{

Currently, \code{plotPars} can include \code{trueValues} generated by \code{mastSim} and \code{SAVEPLOTS}.  If \code{SAVEPLOTS = T}, plots are saved to files in \code{outFolder}.  Otherwise, plots are rendered to the screen.

More detailed vignettes can be obtained with:

\code{browseVignettes('mastif')}
}

\value{

Currently, there are only graphical outputs.
}
\examples{
\dontrun{
# simulate data 
seedNames  <- specNames  <- 'acerRubr'
sim <- list(nyr=10, ntree=30, nplot=5, ntrap=40, meanDist = 15,
            specNames = specNames, seedNames = seedNames)
inputs <- mastSim(sim)
output <- mastif( inputs = inputs, ng = 2000, burnin = 1000 )

plotPars <- list(trueValues = inputs$trueValues)
mastPlot(output, plotPars)
}
}
\author{
James S Clark, \email{jimclark@duke.edu}
}
\references{
Clark, J.S., S. LaDeau, and I. Ibanez. 2004. Fecundity of trees and the colonization-competition hypothesis, Ecological Monographs, 74, 415-442.
}
\seealso{
\code{\link{mastSim}} simulates data

A more detailed vignette is can be obtained with:

\code{browseVignettes('mast')}

website 'http://sites.nicholas.duke.edu/clarklab/code/'.
}

